project fix, stage 2. Fixed warnings, combined two reducers.
