+++ 6/src/actions/searchTitle.js
@@ -1,8 +1,8 @@
-import SEARCH_TITLE from "../constantsAndTypes/types"
+import {SEARCH_TITLE} from "../constantsAndTypes/types"
 
 export default function searchTitle (searchTitle) {
     return {
-        type: "SEARCH_TITLE",
+        type: SEARCH_TITLE,
         searchTitle
     }
 }
