+++ 6/src/common/CommentList.js
@@ -1,11 +1,10 @@
 import React, { Component } from 'react';
-import { connect } from "react-redux";
 import Comment from "./Comment";
 
 export default class CommentList extends Component{
     render(){
-        var commentsToFilm = this.props.comments.map((commentToFilm)=>{
-            return <Comment comment = {commentToFilm} key = {commentToFilm.id}/>        
+        var commentsToFilm = this.props.comments.map((commentToFilm, index)=>{
+            return <Comment comment = {commentToFilm} key = {index}/>        
             }
         )
         return(
