--- 6/src/reducers/orderByReducer.js
@@ -1,11 +1,0 @@
-import { ORDER_BY } from "../constantsAndTypes/types";
-
-export const orderByReducer = (state = "", action) => {
-    switch(action.type){
-        case ORDER_BY:
-          return action.orderBy;
-          break;
-        default:
-          return state;
-    }
-}
