+++ 6/src/reducers/orderSearchReducers.js
@@ -1,4 +1,15 @@
+import { ORDER_BY } from "../constantsAndTypes/types";
 import { SEARCH_TITLE } from "../constantsAndTypes/types";
+
+export const orderByReducer = (state = "", action) => {
+    switch(action.type){
+        case ORDER_BY:
+          return action.orderBy;
+          break;
+        default:
+          return state;
+    }
+}
 
 export const searchTitleReducer = (state = "", action) => {
     switch(action.type){
